requirements.txt
google-generativeai
chromadb
streamlit
python-dotenv